QrCode/requirements.txt
colorama==0.4.4
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.33.3
kiwisolver==1.4.2
numpy==1.22.4
opencv-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
qrcode==7.3.1
six==1.16.0
